+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `s.blankable` Spec for specifying optional string values (#45)
+- Add `s.default` Spec for specifying default values on failing Specs (#46)
 
 
 ## [v0.2.3] - 2019-11-27